% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{cauchy_prior}
\alias{cauchy_prior}
\title{Create a cauchy prior object}
\usage{
cauchy_prior(loc, scale)
}
\arguments{
\item{loc}{location parameter}

\item{scale}{scale parameter (must be > 0)}
}
\value{
A list with dist="cauchy" and values=c(loc, scale)
}
\description{
Create a cauchy prior object
}
