% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodels.R
\name{compile_rubin_normal_mod}
\alias{compile_rubin_normal_mod}
\title{Compile the rubin_normal Stan model (internal)}
\usage{
compile_rubin_normal_mod()
}
\value{
A \code{stanmodel} object corresponding to \code{rubin_normal.stan}.
}
\description{
This function checks if the rubin_normal Stan model is already compiled
(stored in the \code{stanmodels} environment). If not, it compiles
\code{rubin_normal.stan} using \code{rstan::stan_model()} and caches
the resulting model in \code{stanmodels$rubin_normal}.
}
\details{
The \code{rubin_normal.stan} file is assumed to reside in the
\code{inst/stan/} folder of this package. The compiled model is stored
in \code{stanmodels$rubin_normal}, so subsequent calls will not re-compile.

This function is primarily called internally by \code{fit_rubin_normal()}.
Most users will not need to call it directly.
}
\keyword{internal}
