% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{half_cauchy_prior}
\alias{half_cauchy_prior}
\title{Create a half-Cauchy prior for scale parameters}
\usage{
half_cauchy_prior(loc = 0, scale = 5)
}
\arguments{
\item{loc}{Typically 0 if it's strictly positive and centered at 0}

\item{scale}{Must be >0}
}
\value{
A list with dist="half_cauchy" and values=c(loc, scale)
}
\description{
This is another popular choice for scale parameters in hierarchical models,
recommended for example in Gelman (2006).
}
