% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{half_student_t_prior}
\alias{half_student_t_prior}
\title{Create a half-Student-t prior for scale parameters}
\usage{
half_student_t_prior(df, scale)
}
\arguments{
\item{df}{degrees of freedom}

\item{scale}{scale parameter}
}
\value{
A list with dist="half_student_t" and values=c(df, scale, 0)
}
\description{
Often recommended for hierarchical standard deviations like tau (Gelman 2006).
}
\details{
Internally you can treat half-Student-t by bounding a Student-t at 0.
For your Stan code, you might do real<lower=0> tau;  tau ~ student_t(...);
or an explicit T\link{0,} truncation. This function just signals to the code that
it's a half_t.
}
