% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{lkj}
\alias{lkj}
\title{An LKJ prior for correlation matrices}
\usage{
lkj(shape, order = NULL)
}
\arguments{
\item{shape}{shape parameter for the LKJ distribution}

\item{order}{dimension of the correlation matrix (may be optional)}
}
\value{
list with dist="lkj", values=c(shape)
}
\description{
An LKJ prior for correlation matrices
}
