% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{lognormal_prior}
\alias{lognormal_prior}
\title{Create a lognormal prior object}
\usage{
lognormal_prior(mu, sigma)
}
\arguments{
\item{mu}{mean of ln(X) for lognormal}

\item{sigma}{standard deviation of ln(X) for lognormal, must be >0}
}
\value{
A list with dist="lognormal" and values=c(mu, sigma)
}
\description{
Create a lognormal prior object
}
