% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{multinormal}
\alias{multinormal}
\title{A placeholder for multinormal prior}
\usage{
multinormal(location, Sigma)
}
\arguments{
\item{location}{numeric vector of means}

\item{Sigma}{variance-covariance matrix}
}
\value{
A list with dist="multinormal", mean=location, scale=Sigma
}
\description{
This can handle vector hypermeans. Typically you won't use it for scalar mu/tau,
but for multi-dimensional hypermeans if your model is extended.
}
