% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{normal_prior}
\alias{normal_prior}
\title{Create a normal prior object}
\usage{
normal_prior(loc, scale)
}
\arguments{
\item{loc}{Mean (location) of the normal distribution}

\item{scale}{Standard deviation of the normal distribution (must be > 0)}
}
\value{
A list with elements dist="normal" and values=c(loc, scale).
}
\description{
Create a normal prior object
}
