% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_rubin_normal.R
\name{uniform_prior}
\alias{uniform_prior}
\title{Create a uniform prior object}
\usage{
uniform_prior(lower, upper)
}
\arguments{
\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
A list with dist="uniform" and values=c(lower, upper)
}
\description{
Create a uniform prior object
}
